pyproject.toml
[project]
name = "gocortexbrokenbank"
version = "1.3.6"
description = "GoCortex Broken Bank - Intentionally vulnerable banking application for CI/CD security testing and educational purposes"
license = "AGPL-3.0-or-later"
authors = [
    {name = "Simon Sigre", email = "simon.sigre@gocortex.com"},
]
requires-python = ">=3.11"
dependencies = [
    # Vulnerable packages for SAST testing (pinned to vulnerable versions)
    "cryptography==39.0.0",
    "email-validator==1.1.3",
    "flask==2.0.1",
    "flask-login==0.5.0",
    "flask-sqlalchemy==2.5.1",
    "gunicorn==20.0.4",
    "ldap3==2.8.1",
    "psycopg2-binary==2.9.6",
    "pyjwt==2.1.0",
    "pymongo==3.11.4",
    "pyyaml==6.0",
    "requests==2.25.1",
    "sqlalchemy==1.4.23",
    "urllib3==1.26.5",
    "werkzeug==2.0.1",
    "jinja2==3.0.1",
    # PyGremlinBox varied license packages for license compliance testing
    "pygremlinbox-agpl-1-0==1.4.6",
    "pygremlinbox-agpl-1-0-only==1.4.6",
    "pygremlinbox-agpl-1-0-or-later==1.4.6",
    "pygremlinbox-agpl-3-0==1.4.6",
    "pygremlinbox-agpl-3-0-only==1.4.6",
    "pygremlinbox-agpl-3-0-or-later==1.4.6",
    "pygremlinbox-apsl==1.4.6",
    "pygremlinbox-arphic-1999==1.4.6",
    "pygremlinbox-artistic-1-0==1.4.6",
    "pygremlinbox-busl-1-1==1.4.6",
    "pygremlinbox-c-uda-1-0==1.4.6",
    "pygremlinbox-cal-1-0-combined-work-exception==1.4.6",
    "pygremlinbox-cc-by-nc-3-0-de==1.4.6",
    "pygremlinbox-cc-by-nc-nd-3-0-de==1.4.6",
    "pygremlinbox-cc-by-nc-nd-3-0-igo==1.4.6",
    "pygremlinbox-cc-by-nc-sa-2-0-de==1.4.6",
    "pygremlinbox-cc-by-nc-sa-2-0-fr==1.4.6",
    "pygremlinbox-cc-by-nc-sa-2-0-uk==1.4.6",
    "pygremlinbox-cc-by-nc-sa-3-0-de==1.4.6",
    "pygremlinbox-cc-by-nc-sa-3-0-igo==1.4.6",
    "pygremlinbox-cc-by-nd-3-0-de==1.4.6",
    "pygremlinbox-cc-by-sa-2-0-uk==1.4.6",
    "pygremlinbox-cc-by-sa-2-1-jp==1.4.6",
    "pygremlinbox-cc-by-sa-3-0-at==1.4.6",
    "pygremlinbox-cc-by-sa-3-0-de==1.4.6",
    "pygremlinbox-cc-by-sa-4-0==1.4.6",
    "pygremlinbox-cddl-1-0==1.4.6",
    "pygremlinbox-cdla-sharing-1-0==1.4.6",
    "pygremlinbox-cern-ohl-s-2-0==1.4.6",
    "pygremlinbox-cern-ohl-w-2-0==1.4.6",
    "pygremlinbox-copyleft-next-0-3-0==1.4.6",
    "pygremlinbox-copyleft-next-0-3-1==1.4.6",
    "pygremlinbox-cpol-1-02==1.4.6",
    "pygremlinbox-ecos-2-0==1.4.6",
    "pygremlinbox-epl-1-0==1.4.6",
    "pygremlinbox-epl-2-0==1.4.6",
    "pygremlinbox-eupl-1-1==1.4.6",
    "pygremlinbox-eupl-1-2==1.4.6",
    "pygremlinbox-eupl-3-0==1.4.6",
    "pygremlinbox-fdk-aac==1.4.6",
    "pygremlinbox-gpl-2-0==1.4.6",
    "pygremlinbox-gpl-3-0==1.4.6",
    "pygremlinbox-hippocratic-2-1==1.4.6",
    "pygremlinbox-jpl-image==1.4.6",
    "pygremlinbox-lgpl-2-0==1.4.6",
    "pygremlinbox-lgpl-2-1==1.4.6",
    "pygremlinbox-lgpl-3-0==1.4.6",
    "pygremlinbox-linux-man-pages-copyleft==1.4.6",
    "pygremlinbox-mpl-1-1==1.4.6",
    "pygremlinbox-mpl-2-0==1.4.6",
    "pygremlinbox-ms-lpl==1.4.6",
    "pygremlinbox-ncgl-uk-2-0==1.4.6",
    "pygremlinbox-openpbs-2-3==1.4.6",
    "pygremlinbox-osl-3-0==1.4.6",
    "pygremlinbox-polyform-noncommercial-1-0-0==1.4.6",
    "pygremlinbox-polyform-small-business-1-0-0==1.4.6",
    "pygremlinbox-qpl-1-0-inria-2004==1.4.6",
    "pygremlinbox-sendmail-8-23==1.4.6",
    "pygremlinbox-simpl-2-0==1.4.6",
    "pygremlinbox-sspl-1-0==1.4.6",
    "pygremlinbox-tapr-ohl-1-0==1.4.6",
    "pygremlinbox-tpl-1-0==1.4.6",
    "pygremlinbox-ucl-1-0==1.4.6",
    "pygremlinbox-unlicense==1.4.6",
    "pygremlinbox-wxwindows==1.4.6",
    # PyGremlinBox malware simulation packages for security testing
    "pygremlinbox-malware-network-indicators==1.4.6",
    "pygremlinbox-malware-c2-beacon==1.4.6",
    "pygremlinbox-malware-code-obfuscation==1.4.6",
    "pygremlinbox-malware-install-execution==1.4.6",
    "pygremlinbox-malware-credential-harvesting==1.4.6",
    "pygremlinbox-malware-cryptomining-indicators==1.4.6",
]
